requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
PyPDF2>=3.0.1
python-docx>=1.1.0
transformers>=4.30.2
torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
numpy>=1.24.3
python-dotenv>=1.0.0
aiofiles>=23.2.1
huggingface_hub>=0.16.4 